requirements_otp.txt
python-dotenv==1.0.0
requests==2.31.0
twilio==8.11.0
